sync-server-sample/app/requirements.txt
numpy
python-socketio
python-engineio
requests
websocket-client
wget
prometheus-client
